Creating PrincipalNotifications style file
