+++ src/Styles/PrincipalNotificationsStyles.js
@@ -1,0 +1,23 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    principal: {
+      flex: 1,
+      alignItems: 'center',
+      backgroundColor: 'white',
+    },
+  
+    button: {
+      width: 320,
+      height: 65,
+      alignItems: 'center',
+      justifyContent: 'center',
+      backgroundColor: '#FF9500',
+      borderRadius: 10,
+      borderColor: 'black',
+      borderWidth: 1,
+      marginBottom: 60,
+    },
+});
+
+export default styles;
