+++ src/screens/PrincipalNotifications.js
@@ -3,26 +3,7 @@ import { StyleSheet, Text, View, TouchableOpacity, BackHandler } from 'react-nat
 import { Actions } from 'react-native-router-flux';
 import Header from '../components/Header';
 import { backHandlerPopToMain } from '../NavigationFunctions';
-
-const styles = StyleSheet.create({
-  principal: {
-    flex: 1,
-    alignItems: 'center',
-    backgroundColor: 'white',
-  },
-
-  button: {
-    width: 320,
-    height: 65,
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: '#FF9500',
-    borderRadius: 10,
-    borderColor: 'black',
-    borderWidth: 1,
-    marginBottom: 60,
-  },
-});
+import styles from '../Styles/PrincipalNotificationsStyles'
 
 class PrincipalNotifications extends React.PureComponent {
   componentWillMount() {
